refactor dep inference intrinsics
